Add new work.

+++ DIUTask/DIUTask/student/StudentView.cs
@@ -18,15 +18,56 @@ namespace DIUTask.student
         {
             InitializeComponent();
             txtApplyDateMain.Text = System.DateTime.Now.ToString("dd/mm/yyyy"); 
-            txtAddressTmp.Text = System.DateTime.Now.ToString("dd/mm/yyyy");
+            txtApplyDateTmp.Text = System.DateTime.Now.ToString("dd/mm/yyyy");
             lblRegNumber.Text = "123";
             checkTempCartificateAvailable();
             checkMainCartificateAvailable();
         }
 
-        private void label1_Click(object sender, EventArgs e)
+        private bool checkAllDataEantryOrNot(int fromIndex)
         {
+            if (fromIndex == 1)
+            {
+                if(txtRegNumberTmp.Text != String.Empty
+                   && txtNameTmp.Text != String.Empty
+                   && txtFatherNameTmp.Text != String.Empty
+                   && txtAddressTmp.Text != String.Empty
+                   && txtMobileTmp.Text != String.Empty
+                   && txtEmailTmp.Text != String.Empty
+                   && txtApplyDateTmp.Text != String.Empty
+                   && txtTranITTmp.Text != String.Empty
+                   && txtStatusTmp.Text != String.Empty)
+                {
+                    return true;
+                }
+                else
+                {
+                    MessageBox.Show("Please fulfill all requirements first.");
+                }
+            }else if (fromIndex == 2)
+            {
+                if (txtRegNumberMain.Text != String.Empty
+                   && txtNameMain.Text != String.Empty
+                   && txtFatherNameMain.Text != String.Empty
+                   && txtAddressMain.Text != String.Empty
+                   && txtMobileMain.Text != String.Empty
+                   && txtEmailMain.Text != String.Empty
+                   && txtApplyDateMain.Text != String.Empty
+                   && txtTranITMain.Text != String.Empty
+                   && txtStatusMain.Text != String.Empty
+                   && txtSessionMain.Text != String.Empty
+                   && txtProgramMain.Text != String.Empty
+                   && txtCGPAMain.Text != String.Empty)
+                {
+                    return true;
+                }
+                else
+                {
+                    MessageBox.Show("Please fulfill all requirements first.");
+                }
+            }
 
+            return false;
         }
 
         private void tabCartificateStatus_SelectedIndexChanged(object sender, EventArgs e)
@@ -43,39 +84,43 @@ namespace DIUTask.student
         {
             TempCartificateModel model = new TempCartificateModel();
 
-            model.reg_number = txtRegNumberTmp.Text;
-            model.name = txtNameTmp.Text;
-            model.father_name = txtFatherNameTmp.Text;
-            model.address = txtAddressTmp.Text;
-            model.mobile = txtMobileTmp.Text;
-            model.email = txtEmailTmp.Text;
-            model.apply_date = System.DateTime.Now;
-            model.tranit = txtTranITTmp.Text;
-            model.status = txtStatusTmp.Text;
+            if (checkAllDataEantryOrNot(1))
+            {
+                model.reg_number = txtRegNumberTmp.Text;
+                model.name = txtNameTmp.Text;
+                model.father_name = txtFatherNameTmp.Text;
+                model.address = txtAddressTmp.Text;
+                model.mobile = txtMobileTmp.Text;
+                model.email = txtEmailTmp.Text;
+                model.apply_date = System.DateTime.Now;
+                model.tranit = txtTranITTmp.Text;
+                model.status = txtStatusTmp.Text;
 
-            model.insertTempCartificateData();
-
-
+                model.insertTempCartificateData();
+            }
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
             MainCartificateModel model = new MainCartificateModel();
 
-            model.reg_number = txtRegNumberMain.Text;
-            model.name = txtNameMain.Text;
-            model.father_name = txtFatherNameMain.Text;
-            model.address = txtAddressMain.Text;
-            model.mobile = txtMobileMain.Text;
-            model.email = txtEmailMain.Text;
-            model.apply_date = System.DateTime.Now;
-            model.tranit = txtTranITMain.Text;
-            model.status = txtStatusMain.Text;
-            model.session = txtSessionMain.Text;
-            model.program = txtSessionMain.Text;
-            model.CGPA = float.Parse(txtCGPAMain.Text);
+            if (checkAllDataEantryOrNot(2))
+            {
+                model.reg_number = txtRegNumberMain.Text;
+                model.name = txtNameMain.Text;
+                model.father_name = txtFatherNameMain.Text;
+                model.address = txtAddressMain.Text;
+                model.mobile = txtMobileMain.Text;
+                model.email = txtEmailMain.Text;
+                model.apply_date = System.DateTime.Now;
+                model.tranit = txtTranITMain.Text;
+                model.status = txtStatusMain.Text;
+                model.session = txtSessionMain.Text;
+                model.program = txtSessionMain.Text;
+                model.CGPA = float.Parse(txtCGPAMain.Text);
 
-            model.insertMainCartificateData();
+                model.insertMainCartificateData();
+            }
         }
 
         private void checkTempCartificateAvailable()
@@ -120,9 +165,41 @@ namespace DIUTask.student
             {
                 File.Delete(downloadPath);
             }
-            File.Move(uploadedPath, downloadPath);
+            if (File.Exists(uploadedPath))
+            {
+                File.Move(uploadedPath, downloadPath);
+                MessageBox.Show("File download in " + downloadPath);
+            }
+            else
+            {
+                MessageBox.Show("File not exists !");
+            }
+            
+        }
 
-            MessageBox.Show("File download in "+downloadPath);
+        private void btnMainDownload_Click(object sender, EventArgs e)
+        {
+            string uploadedPath=new MainCartificateModel().getFileLocation(lblRegNumber.Text);
+            string[] paths = uploadedPath.Split('\\');
+
+            folderBrowserDialogBox.ShowDialog();
+
+            string downloadPath = folderBrowserDialogBox.SelectedPath + "\\" + paths[paths.Length - 1];
+
+            if (File.Exists(downloadPath))
+            {
+                File.Delete(downloadPath);
+            }
+            if (File.Exists(uploadedPath))
+            {
+                File.Move(uploadedPath, downloadPath);
+                MessageBox.Show("File download in " + downloadPath);
+            }
+            else
+            {
+                MessageBox.Show("File not exists !");
+            }
+            
         }
     }
 }
